requirements.txt
intel-tensorflow==2.7.0
keras
numpy
matplotlib
pillow
pydot
